refactor(api.response): move file and edit existing logic to add it
