+++ src/middlewares/permission.middleware.ts
@@ -1,17 +1,18 @@
 import type { NextFunction, Response } from 'express';
 import mongoose from 'mongoose';
 
+import { Errors } from '../environments/errors.environment';
 import type { IRequest } from '../types/global.type';
-import { errorFormatter, ErrorMessages } from '../utils/errors';
+import { ApiResponse } from '../utils/apiResponse';
 import { Roles } from '../utils/roles';
 
 export const isInGroup = (rolesGranted: number[]) => (req: IRequest, res: Response, next: NextFunction) => {
-  if (!req.user?.role) return errorFormatter(res, 401, ErrorMessages.UNAUTHORIZED);
+  if (!req.user?.role) return new ApiResponse(res, Errors.UNAUTHORIZED_RESPONSE);
 
   const roles = [...rolesGranted];
 
   const isGranted = roles.includes(req.user?.role);
-  if (!isGranted) return errorFormatter(res, 403, ErrorMessages.FORBIDDEN);
+  if (!isGranted) return new ApiResponse(res, Errors.FORBIDDEN_RESPONSE);
 
   return next();
 };
@@ -25,24 +26,24 @@ export const checkOwnership =
       const role = req.user?.role;
       const isValidObjectId = mongoose.Types.ObjectId.isValid(dataId);
       if (!isValidObjectId) {
-        return errorFormatter(res, 400, ErrorMessages.MALFORMED_DATA);
+        return new ApiResponse(res, Errors.BAD_REQUEST_RESPONSE);
       }
 
       const model = mongoose.model(modelName);
       const doc = await model.findById(dataId);
 
       if (!doc) {
-        return errorFormatter(res, 404, ErrorMessages.NOT_FOUND);
+        return new ApiResponse(res, Errors.NOT_FOUND_RESPONSE);
       }
 
       const ownerId = doc.get(ownerIdPath);
 
       if (userId !== ownerId.toString() && role !== Roles.ADMIN) {
-        return errorFormatter(res, 403, ErrorMessages.NOT_AUTHORIZED);
+        return new ApiResponse(res, Errors.FORBIDDEN_RESPONSE);
       }
 
       return next();
     } catch (err) {
-      return errorFormatter(res, 500, ErrorMessages.SERVER_ERROR, err);
+      return new ApiResponse(res, Errors.INTERNAL_SERVER_RESPONSE, err);
     }
   };
