--- src/utils/errors.ts
@@ -1,27 +1,0 @@
-import type { Response } from 'express';
-
-export const ErrorMessages = {
-  MALFORMED_DATA: 'Request malformed.',
-  SERVER_ERROR: 'Something went wrong.',
-  CONNECTION_ERROR: 'Error during login.',
-  NOT_FOUND: 'Data not found.',
-  NOT_AUTHORIZED: 'You are not authorized to access to this data.',
-  UNAUTHORIZED: 'Authentification required.',
-  FORBIDDEN: 'Access denied.',
-  ALREADY_EXIST: 'Data already exist.',
-  TOKEN_ERROR: 'Error with access token.',
-};
-
-export const errorFormatter = (res: Response, code: number, message: string, error?: any): Response => {
-  if (process.env.NODE_ENV === 'production') {
-    return res.status(code).json({
-      error: true,
-      message,
-    });
-  }
-  return res.status(code).json({
-    error: true,
-    message,
-    debug: error,
-  });
-};
